+++ pyinaturalist/request_params.py
@@ -29,7 +29,7 @@ from pyinaturalist.converters import (
     try_int,
 )
 
-COMMON_PARAMS = ['access_token', 'dry_run', 'user_agent', 'session']
+COMMON_PARAMS = ['access_token', 'dry_run', 'session']
 MULTIPLE_CHOICE_ERROR_MSG = (
     'Parameter "{}" must have one of the following values: {}\n\tValue provided: {}'
 )
@@ -44,10 +44,10 @@ INTERVALS: Dict[str, Union[timedelta, relativedelta]] = {
 logger = getLogger(__name__)
 
 
-def preprocess_request_body(body: Optional[RequestParams]) -> RequestParams:
+def preprocess_request_body(body: Optional[RequestParams]) -> Optional[RequestParams]:
     """Perform type conversions, sanity checks, etc. on JSON-formatted request body"""
     if not body:
-        return {}
+        return None
     if 'observation' in body:
         body['observation'] = preprocess_request_params(body['observation'])
     else: