+++ HISTORY.md
@@ -1,5 +1,9 @@
 # History
 
+## 0.15.1 (2021-TBD)
+* Remove `pyinaturalist.user_agent` global variable and API function keyword args, and recommend
+  setting on session object instead
+
 ## 0.15.0 (2021-09-07)
 
 ### New Endpoints